requirements_specific_versions.txt
python-dotenv==1.0.1
download_financial_statement
langchain==0.1.20
AgentType==0.1.20 
ChatOpenAI==0.0.38
create_csv_agent
ChatOpenAI
OpenAI==1.34.0 
SystemMessage==0.1.20
RailsConfig==0.9.0
LLMRails==0.9.0
os==(Python Standard Library)
pandas==2.2.2
requests==2.32.3
NIM_verifier
warnings==(Python Standard Library)